requirements.txt
streamlit==1.38.0
sounddevice==0.4.7
numpy==1.26.4
python-dotenv==1.0.1
groq==0.11.0
openai==1.37.0
httpx==0.27.2
httpcore==1.0.4
